Bug 2009552 - avoid integer overflow in platform-independent ghash. r=nss-reviewers,nkulatova
Differential Revision: https://phabricator.services.mozilla.com/D278681

--HG--
branch : NSS_3_121_BRANCH
extra : source : 245385e16fa62111d6e3c3fbd847b020755f64f0

+++ lib/freebl/gcm.c
@@ -436,7 +436,7 @@ gcmHash_Update(gcmHashContext *ghash, const unsigned char *buf,
     unsigned int blocks;
     SECStatus rv;
 
-    ghash->cLen += (len * PR_BITS_PER_BYTE);
+    ghash->cLen += ((uint64_t)len * PR_BITS_PER_BYTE);
 
     /* first deal with the current buffer of data. Try to fill it out so
      * we can hash it */
